+++ Panther/src/symbol_proc/SymbolProcBuilder.cpp
@@ -2622,12 +2622,27 @@ namespace pcit::panther{
 			} break;
 
 			case Token::lookupKind("||"): case Token::lookupKind("&&"): {
-				this->emit_error(
-					Diagnostic::Code::MISC_UNIMPLEMENTED_FEATURE,
-					node,
-					"Building symbol proc of [&&] and [||] is unimplemented"
-				);
-				return evo::resultError;
+				const evo::Result<SymbolProc::TermInfoID> lhs = this->analyze_expr<IS_CONSTEXPR>(infix.lhs);
+				if(lhs.isError()){ return evo::resultError; }
+
+				const evo::Result<SymbolProc::TermInfoID> rhs = this->analyze_expr<IS_CONSTEXPR>(infix.rhs);
+				if(rhs.isError()){ return evo::resultError; }
+
+				const SymbolProc::TermInfoID new_term_info_id = this->create_term_info();
+				if constexpr(IS_CONSTEXPR){
+					this->add_instruction(
+						this->context.symbol_proc_manager.createMathInfixConstexprLogical(
+							infix, lhs.value(), rhs.value(), new_term_info_id
+						)
+					);
+				}else{
+					this->add_instruction(
+						this->context.symbol_proc_manager.createMathInfixLogical(
+							infix, lhs.value(), rhs.value(), new_term_info_id
+						)
+					);
+				}
+				return new_term_info_id;
 			} break;